use if to break infinite loop level 1

+++ image/root/src/bin/uniquespace-docker-container-prune-deprecated.sh
@@ -2,7 +2,9 @@
 
 docker container ls --quiet --all --filter label=commit-id | while read CONTAINER
 do
-    [ "${COMMIT_ID}" != "$(docker container inspect --format "{{index .Config.Labels \"commit-id\"}}" ${CONTAINER})" ] &&
+    if [ "${COMMIT_ID}" != "$(docker container inspect --format "{{index .Config.Labels \"commit-id\"}}" ${CONTAINER})" ]
+    then
         docker container stop ${CONTAINER} &&
         docker container rm --volumes ${CONTAINER}
+    fi
 done
