+++ image/root/src/bin/uniquespace-docker-container-prune-expired.sh
@@ -2,7 +2,9 @@
 
 docker container ls --quiet --all --filter label=commit-id | while read CONTAINER
 do
-    [ "$(date +%s)" -gt "$(docker container inspect --format "{{index .Config.Labels \"expiry\"}}" ${CONTAINER})" ] &&
+    if [ "$(date +%s)" -gt "$(docker container inspect --format "{{index .Config.Labels \"expiry\"}}" ${CONTAINER})" ]
+    then
         docker container stop ${CONTAINER} &&
         docker container rm --volumes ${CONTAINER}
+    fi
 done
