+++ image/root/src/bin/uniquespace-docker-volume-prune-deprecated.sh
@@ -2,6 +2,8 @@
 
 docker volume ls --quiet --filter label=commit-id | while read VOLUME
 do
-    [ "${COMMIT_ID}" != "$(docker volume inspect --format "{{index .Labels \"commit-id\"}}" ${VOLUME})" ] &&
+    if [ "${COMMIT_ID}" != "$(docker volume inspect --format "{{index .Labels \"commit-id\"}}" ${VOLUME})" ]
+    then
         docker volume rm ${VOLUME}
+    fi
 done
